Update tests + lint
